[MC-1508]Make the NameRecord a seperate module
