+++ app/controllers/JsonController.scala
@@ -5,7 +5,7 @@ import play.api.libs.json._
 
 import play.api.libs.json._
 import play.api.libs.functional.syntax._
-case class NameRecord(name: String, age: Int)
+import module._
 
 class JsonController extends Controller{
   val parsedJsonValue = Json.parse("""
@@ -28,7 +28,7 @@ class JsonController extends Controller{
   }
   
   var nameList :List[NameRecord] = Nil
-  def jsString :String = Json.toJson(nameList).toString
+  def jsString :String = NameRecord.listToString(nameList)
   
   def addName(name: String,age: Int) = Action{request =>
     val record = NameRecord(name,age)
@@ -38,33 +38,12 @@ class JsonController extends Controller{
   }
   
   def parseStr = Action{request =>
-    val json = Json.parse(jsString)
-    val jsonList = json.validate[List[NameRecord]]
     
-    jsonList match {
-      case success: JsSuccess[List[NameRecord]] => 
-      {
-        val myList = success.get
-        myList.foreach(jsrecord => {
-          println("name :" + jsrecord.name + " age:" + jsrecord.age)
-        })
-
-      }
-      case e: JsError => println("Errors")
-    }
     
-    Ok("ok")
+    Ok(NameRecord.parseJsonValueIntoNameRecord(jsString))
       
   }
   
-  implicit val nameRecordWrite: Writes[NameRecord] = (
-      (JsPath \ "name").write[String] and
-      (JsPath \ "age").write[Int]
-  )(unlift(NameRecord.unapply))
   
-  implicit val nameRecordRead: Reads[NameRecord] = (
-    (JsPath \ "name").read[String] and
-    (JsPath \ "age").read[Int]
-  )(NameRecord.apply _)
 
 }
